move banking tasks to own file
