+++ utils.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 # coding=utf-8
 from datetime import datetime
-import ofxparse
 import json
 from collections import OrderedDict
 
@@ -14,191 +13,25 @@ from sqlalchemy.orm.exc import NoResultFound
 
 from main import create_app, mail, db
 from models import (
-    User, TicketType, Ticket, TicketPrice, TicketLimitException
-)
-from models.payment import (
-    BankAccount, BankTransaction,
+    User, TicketType, Ticket, TicketPrice
 )
 from models.permission import Permission
 from models.email import EmailJobRecipient
 from models.ical import CalendarSource
-from apps.payments import banktransfer
 from apps.common.receipt import attach_tickets
 
+from utils.banking import (
+    CreateBankAccounts, LoadOfx, Reconcile,
+)
 from utils.cfp import (
     ImportCFP, LockProposals, EmailSpeakersAboutSlot,
     EmailSpeakersAboutFinalising, RejectUnacceptedTalks
 )
-from utils.dev import MakeFakeUsers, MakeFakeTickets
+from utils.dev import CreateDB, MakeFakeUsers, MakeFakeTickets
 from utils.schedule import (
     ImportVenues, SetRoughDurations, OutputSchedulerData, ImportSchedulerData,
     RunScheduler, ApplyPotentialSchedule
 )
-
-class CreateBankAccounts(Command):
-    def run(self):
-        gbp = BankAccount('492900', '20716473590526', 'GBP')
-        eur = BankAccount('492900', '20716472954433', 'EUR')
-        for acct in [gbp, eur]:
-            try:
-                BankAccount.query.filter_by(acct_id=acct.acct_id, sort_code=acct.sort_code).one()
-            except NoResultFound:
-                app.logger.info('Adding %s account %s %s', acct.currency, acct.sort_code, acct.acct_id)
-                db.session.add(acct)
-
-        db.session.commit()
-
-
-class LoadOfx(Command):
-    option_list = [Option('-f', '--file', dest='filename', help="The .ofx file to load")]
-
-    def run(self, filename):
-        ofx = ofxparse.OfxParser.parse(open(filename))
-
-        acct_id = ofx.account.account_id
-        sort_code = ofx.account.routing_number
-        account = BankAccount.get(sort_code, acct_id)
-        if ofx.account.statement.currency.lower() != account.currency.lower():
-            app.logger.error("Currency %s doesn't match account currency %s",
-                             ofx.account.statement.currency, account.currency)
-            return
-
-        added = 0
-        duplicate = 0
-        dubious = 0
-
-        for txn in ofx.account.statement.transactions:
-            if 0 < int(txn.id) < 200101010000000:
-                app.logger.debug('Ignoring uncleared transaction %s', txn.id)
-                continue
-            # date is actually dtposted and is a datetime
-            if txn.date < datetime(2015, 1, 1):
-                app.logger.debug('Ignoring historic transaction from %s', txn.date)
-                continue
-            if txn.amount <= 0:
-                app.logger.info('Ignoring non-credit transaction for %s', txn.amount)
-                continue
-
-            dbtxn = BankTransaction(
-                account_id=account.id,
-                posted=txn.date,
-                type=txn.type,
-                amount=txn.amount,
-                payee=txn.payee,
-                fit_id=txn.id,
-            )
-
-            # Check for matching/duplicate transactions.
-            # Insert if possible - conflicts can be sorted out within the app.
-            matches = dbtxn.get_matching()
-
-            # Euro payments have a blank fit_id
-            if dbtxn.fit_id == '00000000':
-                # There seems to be a serial in the payee field. Assume that's enough for uniqueness.
-                if matches.count():
-                    app.logger.debug('Ignoring duplicate transaction from %s', dbtxn.payee)
-                    duplicate += 1
-
-                else:
-                    db.session.add(dbtxn)
-                    added += 1
-
-            else:
-                different_fit_ids = matches.filter( BankTransaction.fit_id != dbtxn.fit_id )
-                same_fit_ids = matches.filter( BankTransaction.fit_id == dbtxn.fit_id )
-
-                if same_fit_ids.count():
-                    app.logger.debug('Ignoring duplicate transaction %s', dbtxn.fit_id)
-                    duplicate += 1
-
-                elif BankTransaction.query.filter( BankTransaction.fit_id == dbtxn.fit_id ).count():
-                    app.logger.error('Non-matching transactions with same fit_id %s', dbtxn.fit_id)
-                    dubious += 1
-
-                elif different_fit_ids.count():
-                    app.logger.warn('%s matching transactions with different fit_ids for %s',
-                                    different_fit_ids.count(), dbtxn.fit_id)
-                    # fit_id may have been changed, so add it anyway
-                    db.session.add(dbtxn)
-                    added += 1
-                    dubious += 1
-
-                else:
-                    db.session.add(dbtxn)
-                    added += 1
-
-        db.session.commit()
-        app.logger.info('Import complete: %s new, %s duplicate, %s dubious',
-                        added, duplicate, dubious)
-
-
-class Reconcile(Command):
-
-    option_list = [Option('-d', '--doit', action='store_true', help="set this to actually change the db")]
-
-    def run(self, doit):
-        txns = BankTransaction.query.filter_by(payment_id=None, suppressed=False)
-
-        paid = 0
-        failed = 0
-
-        for txn in txns:
-            if txn.type.lower() not in ('other', 'directdep'):
-                raise ValueError('Unexpected transaction type for %s: %s', txn.id, txn.type)
-
-            if txn.payee.startswith("GOCARDLESS ") or txn.payee.startswith("GC C1 EMF"):
-                app.logger.info('Suppressing GoCardless transfer %s', txn.id)
-                if doit:
-                    txn.suppressed = True
-                    db.session.commit()
-                continue
-
-            if txn.payee.startswith("STRIPE PAYMENTS EU ") or txn.payee.startswith("STRIPE STRIPE"):
-                app.logger.info('Suppressing Stripe transfer %s', txn.id)
-                if doit:
-                    txn.suppressed = True
-                    db.session.commit()
-                continue
-
-            app.logger.info("Processing txn %s: %s", txn.id, txn.payee)
-
-            payment = txn.match_payment()
-            if not payment:
-                app.logger.warn("Could not match payee, skipping")
-                failed += 1
-                continue
-
-            app.logger.info("Matched to payment %s by %s for %s %s",
-                payment.id, payment.user.name, payment.amount, payment.currency)
-
-            if txn.amount != payment.amount:
-                app.logger.warn("Transaction amount %s doesn't match %s, skipping",
-                                txn.amount, payment.amount)
-                failed += 1
-                continue
-
-            if txn.account.currency != payment.currency:
-                app.logger.warn("Transaction currency %s doesn't match %s, skipping",
-                                txn.account.currency, payment.currency)
-                failed += 1
-                continue
-
-            if payment.state == 'paid':
-                app.logger.error("Payment %s has already been paid", payment.id)
-                failed += 1
-                continue
-
-            if doit:
-                txn.payment = payment
-                payment.paid()
-                db.session.commit()
-
-                banktransfer.send_confirmation(payment)
-
-            app.logger.info("Payment reconciled")
-            paid += 1
-
-        app.logger.info('Reconciliation complete: %s paid, %s failed', paid, failed)
 
 def add_ticket_types(types):
     for tt in types:
